Fix detection of tcl and python with -Wl,--as-needed
The problem here was that libs where put into $LDFLAGS where they don't belong.
When you also use -Wl,--as-needed, sth like this is called:

  gcc -Wl,--as-needed -ltcl8.4 your.o objects.o here.o

Due to --as-needed, the linker will drop libtcl8.4 because by the time it sees
that option, nothing depends on it. And if nothing depends on some lib, it can
be dropped.

The fix here is obviously to put -ltcl8.4 into $LIBS (which is appended, not
prepended, to the command line).

The same reasoning goes for python.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@2198 726aef4b-f618-498e-8847-2d620e286838