Pass AF_UNSPEC to ares when the socket address says so.
Ares only attempts a A lookup after a failed AAAA lookup if AF_UNSPEC was
specified in the original request.

This resolves an issue with the colloquy push plug-in on Mac OS X Lion
where 'colloquy.mobi' failed to be resolved (it does not have a AAAA
record) when znc is compiled with IPv6 support.