antiidle: Don't SIGABRT when receiving a line with less than 2 words
This was introduced in 232d2612fea1230681d9c55d74eaab09b273206f which
tries to split a line by spaces and then tries splitted[1] and
splitted[2] which will cause a SIGABRT if the line isn't actually that
long.

This also uses the 4th argument to the 301 line to determine if this
message is because we sent a message to ourselfs. Previously the 3rd
option was used, which is always our own nick. This was resulting in any
301 line being halted.