Fix AX_PTHREAD to actually work
At first the test was failing with the following error due to my CXXFLAGS:

conftest.cpp:20:37: error: parameter 'a' set but not used
[-Werror=unused-but-set-parameter]

Then it figured out that -lpthread works, but I wondered why the test for
"-pthread" still fails. This is where I realized that it is testing with C++
compiler, but puts its stuff in $CFLAGS which is just ignored. So a little
search-and-replace later, this now uses $CXXFLAGS.

At this point the results of the tests looked good, but znc failed to link,
because it didn't use -pthread when linking. One little "PTHREAD_LIBS=" later,
that problem was also gone and I can finally commit this. Yay.

Signed-off-by: Uli Schlachter <psychon@znc.in>