Display a 401 when a partyline user is not attached
Send private partyline messages to each client seperately using the
correct nickname for that client. It may differ if they are user clients
or network clients where they are not connected to an IRC server.

I have created macro's for the nickname prefix so this could easily be
changed.

Use 401 instead of 403 for user/channel not found

Fixes #21