imapauth: Follow RFC more closely
The IMAP RFC allows the server to announce its capabilities before replying to
the LOGIN command. imapauth would misinterpret that as a failed login.

The fix is to only handle lines which contain the tag ("AUTH") used for the
login command.

Thanks to rlpowell for reporting that imapauth doesn't work against
imap.google.com and for testing the fix.

Signed-off-by: Uli Schlachter <psychon@znc.in>