systemWideConfig implementation
This patch adds a systemWideConfig mode to znc.

-S <user>
	enables systemWideConfig
	switches to <user> if znc was run as root

long version of -S: --system-wide-config-as

When enabled, if running as root, znc will drop privileges to the
specified user.

WriteNewConfig has also been modified to skip the "Launch ZNC now" step
when systemWideConfig is enabled.

Per my initial talk with psychon, the new parameter is not documented
in -h for now.

The patch was improved with the help of DGandalf