Makefile: Fix generation of the VERSION_EXTRA flag
When you checkout the znc-1.0 tag and try to build znc, this will fail. The
reason is that "git describe" prints just "znc-1.0" and the sed-magic in the
Makefile fails to handle this correctly.

Fix this by moving this magic into a short shell script which doesn't try to
parse the output of "git describe" but instead gets its information from various
sources directly.

Signed-off-by: Uli Schlachter <psychon@znc.in>