Fix a random crash with module hooks
How to reproduce:
1. have a global module which in a hook does something which calls
hooks too, e.g. Broadcast()
2. have another global module loaded after the first one which uses
m_pNetwork in that hook
3. enjoy the crash (or just weird behavior if you're unlucky)

So, when ZNC is setup with the "proper" global modules and in a special
order, one can crash it just from IRC, by triggering the needed module hooks.

Affected version: 1.0

Thanks to ChauffeR (Simone Esposito) for reporting the issue
and helping to debug it.