Always buffer private messages for replay
This is useful if connecting with multiple hosts, and want private
messages to get replayed when connecting a second(+) host, or
reconnecting the same host. To facilitate clearing the buffer, the
ClearAllBuffers command will now also clear the query buffer.

I've also updated the clearbufferonmsg module to clear query buffers on
msg as that seemed appropriate (and how I want things to work).

This is largely based on the work found at
https://gist.github.com/jreese/1206300

Tested with 1.2 release.