webadmin: combine "List Users" & "Add User"
Let add users via the list of users in the similar manner it’s
done for networks. It’s evidently an extra step to add a user,
but on the other hand it gives a nice overview of the users
when adding new ones, and leads to less clutter in the sidebar.