Reorganize "Installing" section slightly
Originally at the time of editing, the "Installation is done with the ..." command blurb comes *before* the notice about if you're building from git.  I believe it is more sane to have that come first in the README section for installing, as it will be seen before the actual installation commands are, so people will first see that if they are building from git they'll need to run `./autogen.sh`.